refactor(core): refactor apigateway.core
