+++ src/dashboard/apigateway/apigateway/tests/core/test_constants.py
@@ -127,18 +127,3 @@ class TestConstants:
             assert result
         else:
             assert not result
-
-    @pytest.mark.parametrize(
-        "value, match",
-        [
-            ("1.0.0", True),
-            ("1.0.0-beta.1", True),
-            ("v1.0.0", False),
-        ],
-    )
-    def test_semver_pattern(self, value, match):
-        result = constants.SEMVER_PATTERN.match(value)
-        if match:
-            assert result
-        else:
-            assert not result
